lock window and refact scrub window
